server/requirements.txt
fastapi>=0.115.9,<1
uvicorn[standard]>=0.34.0,<1
typing-extensions>=4.12.2,<5
orjson>=3.10.15,<4
asyncpg>=0.30.0,<1
sonyflake-py>=1.3.0,<2
Celery[librabbitmq,redis,msgpack]>=5.4.0,<6
transformers[torch]>=4.49.0,<5
torchaudio>=2.6.0,<3
torchvision>=0.21.0,<1